0.4.0
When supplied an appropriately-formatted key via the argument variableKey, both extract_blanks() and quick_analyze() will check values of one data column against the key and replace those values accordingly.

Also removed conditional calling of auto_rename(); instead it is called by default, and conditionals inside the function check if renaming is necessary, and if not it returns the input data unchanged.