Changed value reassign behaviour
In auto_rename(); when column values are replaced, the old values are copied to a new column with name "<old_column>_old" first.

Also fixed some minor printing/grammar problems.

Noticed a new bug which remains unfixed: when replacing values, if a replacement value is used more than once for different original values, the list of replacements is reported incorrectly. I'm not in the mood to fix this but an issue will be lodged for later.